+++ neo4j/src/test/scala/org/meerkat/neo4j/static_analysis/SimpleStaticAnalysisAliasTest.scala
@@ -2,8 +2,25 @@ package org.meerkat.neo4j.static_analysis
 
 import org.neo4j.graphdb.{GraphDatabaseService, Label, Node}
 
-class SimpleStaticAnalysisTest extends StaticAnalysisTest {
+class SimpleStaticAnalysisAliasTest extends StaticAnalysisAliasTest {
+
   override def fillDb(db: GraphDatabaseService): Unit = {
+  // Program on C is
+  //void main() {
+  //    int *a, *b;
+  //    int *c, *d, *e;
+  //    b = a;
+  //    c = &b;
+  //    *d = *c;
+  //    *c = e;
+  //}
+
+
+  def createEdgeWithReversed(f: Node, t: Node, l: String) = {
+      f.createRelationshipTo(t, () => l)
+      t.createRelationshipTo(f, () => "n" + l)
+    }
+
     val a = db.createNode(Label.label("a"))
     val b = db.createNode(Label.label("b"))
     val c = db.createNode(Label.label("c"))
@@ -15,10 +32,6 @@ class SimpleStaticAnalysisTest extends StaticAnalysisTest {
     val `*c` = db.createNode(Label.label("*c"))
     val `*d` = db.createNode(Label.label("*d"))
 
-    def createEdgeWithReversed(f: Node, t: Node, l: String) = {
-      f.createRelationshipTo(t, () => l)
-      t.createRelationshipTo(f, () => "n" + l)
-    }
 
     createEdgeWithReversed(a, b, "a")
     createEdgeWithReversed(`&b`, b, "d")
